+++ solstice/src/main/java/dev/equo/ide/BuildPluginIdeMain.java
@@ -17,6 +17,7 @@ import com.diffplug.common.swt.os.OS;
 import dev.equo.solstice.NestedJars;
 import dev.equo.solstice.SerializableMisc;
 import dev.equo.solstice.ShimIdeBootstrapServices;
+import dev.equo.solstice.SignedJars;
 import dev.equo.solstice.Solstice;
 import dev.equo.solstice.SolsticeManifest;
 import dev.equo.solstice.p2.WorkspaceRegistry;
@@ -88,10 +89,12 @@ public class BuildPluginIdeMain {
 			Objects.requireNonNull(cleanFlag);
 
 			var classpathSorted = Launcher.copyAndSortClasspath(classpath);
+			SignedJars.stripIfNecessary(classpathSorted);
 			var nestedJarFolder = new File(workspaceDir, NestedJars.DIR);
 			for (var nested : NestedJars.inFiles(classpathSorted).extractAllNestedJars(nestedJarFolder)) {
 				classpathSorted.add(nested.getValue());
 			}
+			SignedJars.stripIfNecessary(classpathSorted);
 
 			if (lockFile.hasClasspath() && !classpathSorted.equals(lockFile.readClasspath())) {
 				System.out.println("WARNING! The classpath has changed since this IDE was setup.");
@@ -274,6 +277,7 @@ public class BuildPluginIdeMain {
 
 		var props = new LinkedHashMap<String, String>();
 		props.put("gosh.args", "--quiet --noshutdown");
+		props.put("osgi.nl", "en_US");
 		props.put(Constants.FRAMEWORK_STORAGE_CLEAN, Constants.FRAMEWORK_STORAGE_CLEAN_ONFIRSTINIT);
 		props.put(
 				EquinoxLocations.PROP_INSTANCE_AREA, new File(installDir, "instance").getAbsolutePath());