+++ solstice/src/main/java/dev/equo/ide/IdeHookM2E.java
@@ -1,0 +1,30 @@
+/*******************************************************************************
+ * Copyright (c) 2022 EquoTech, Inc. and others.
+ *
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     EquoTech, Inc. - initial API and implementation
+ *******************************************************************************/
+package dev.equo.ide;
+
+import java.io.File;
+
+@SuppressWarnings("serial")
+public class IdeHookM2E implements IdeHook {
+	final File rootDir;
+
+	public IdeHookM2E(File rootDir) {
+		this.rootDir = rootDir;
+	}
+
+	@Override
+	public IdeHookInstantiated instantiate() throws Exception {
+		return IdeHook.usingReflection("dev.equo.ide.M2EImpl", this);
+	}
+}